requirements.txt
cryptography==3.4.8
pandas==1.3.4
requests==2.26.0
snowflake_sqlalchemy==1.3.3
SQLAlchemy==1.4.22
